Cleanup for initial impl
