+++ as/plugins/org.jboss.ide.eclipse.as.core/jbosscore/org/jboss/ide/eclipse/as/core/eap/xp/ProjectUtils.java
@@ -26,96 +26,6 @@ import org.eclipse.m2e.core.project.IMavenProjectFacade;
 import org.jboss.tools.common.jdt.core.buildpath.ClasspathContainersHelper;
 
 public class ProjectUtils {
-	/*
-	public static Set<?> findInstalledExtensions(Object currentProject) {
-		try {
-			if (currentProject != null && currentProject instanceof IProject) {
-				IResource resource = ((IProject)currentProject).findMember("pom.xml");
-				if (resource != null) {
-				    IPath path = resource.getRawLocation().removeLastSegments(1);
-				    File file = new File(path.toOSString());
-				    ProjectWriter projectWriter = new FileProjectWriter(file);
-				    BuildFile buildFile = new MavenBuildFile(projectWriter);
-				    ListExtensions listExtensions = new ListExtensions(buildFile);
-				    Map<?,?> extensions = (Map<?,?>)getFindInstalledMethod().invoke(listExtensions);
-				    return extensions.keySet(); 
-				}
-			}
-		} catch (IOException | InvocationTargetException | IllegalAccessException e) {
-			e.printStackTrace();
-		}
-		return new HashSet<Object>();
-	}
-	
-	public static void createProject(			
-			String name, 
-			String location,
-			String groupId, 
-			String artefactId, 
-			String version, 
-			String className,
-			HashMap<String, Object> context) {
-		try {
-			if (context == null) {
-				context = new HashMap<String, Object>();
-			}
-			File workspaceFolder = ResourcesPlugin.getWorkspace().getRoot().getLocation().toFile();
-			File projectFolder = new File(location);
-			ProjectWriter projectWriter = new FileProjectWriter(projectFolder);
-			new CreateProject(projectWriter)
-					.groupId(groupId)
-					.artifactId(artefactId)
-					.version(version)
-					.className(className)
-					.doCreateProject(context);
-			Set<MavenProjectInfo> projectSet = null;
-			IProjectConfigurationManager projectConfigurationManager = MavenPlugin.getProjectConfigurationManager();
-			MavenModelManager mavenModelManager = MavenPlugin.getMavenModelManager();
-			LocalProjectScanner scanner = new LocalProjectScanner(
-					workspaceFolder, //
-					projectFolder.getCanonicalPath(), 
-					false, 
-					mavenModelManager);
-			scanner.run(new NullProgressMonitor());
-			projectSet = projectConfigurationManager.collectProjects(scanner.getProjects());
-			ProjectImportConfiguration configuration = new ProjectImportConfiguration();
-			projectConfigurationManager.importProjects(projectSet,
-					configuration, new NullProgressMonitor());
-		} catch (IOException | InterruptedException | CoreException e) {
-			e.printStackTrace();
-		}
-	}
-
-	private static Method getFindInstalledMethod() {	
-		Method result = null;
-		try {
-			result = ListExtensions.class.getDeclaredMethod("findInstalled");
-			result.setAccessible(true);
-		} catch (NoSuchMethodException | SecurityException e) {
-			throw new RuntimeException(e);
-		}
-		return result;
-	}
-
-	public static void installExtension(Object currentProject, Extension extension) {
-		try {
-			if (currentProject instanceof IProject) {
-				IProject project = (IProject) currentProject;
-				ToolSupport support = getToolSupport(project);
-				ToolContext context = new DefaultToolContext(project.getName() + "__installExtension", project, Collections.emptyMap(), Collections.singletonList(extension.getArtifactId()));
-				support.addExtension(context);
-				project.refreshLocal(IResource.DEPTH_INFINITE, new NullProgressMonitor());
-			}		
-		} catch (CoreException e) {
-			throw new RuntimeException(e);
-		}
-	}
-	
-	public static boolean isApplicationYAML(IFile file) {
-		return isQuarkusProject(file.getProject()) &&
-				("application.yaml".equals(file.getName()) || "application.yml".equals(file.getName()));
-	}
-*/
 	public static Object getSelectedProject(Object selectedElement) {
 		Object result = null;
 		if (selectedElement instanceof IResource) {
@@ -151,18 +61,18 @@ public class ProjectUtils {
 	}
 	
 	public static boolean isJBossXpProject(IProject project) {
-		return isJavaProject(project) && isJBossXpProject(JavaCore.create(project));
+		return isJavaProject(project) && isMavenProject(project) && isJBossXpProject(JavaCore.create(project));
 	}
 
 	public static boolean isJBossXpProject(IJavaProject javaProject) {
 		IMavenProjectFacade facade = MavenPlugin.getMavenProjectRegistry().create( javaProject.getProject(), new NullProgressMonitor() );
 	    try {
+	    	if( facade == null ) {
+	    		return false;
+	    	}
 			Plugin plugin = facade.getMavenProject(new NullProgressMonitor()).getPlugin("org.wildfly.plugins:wildfly-jar-maven-plugin");
 			if (plugin != null) {
-	    		//JBossServerUIPlugin.log("Found wf plugin " + plugin, null);
 	    		return true;
-			} else {
-	    		//JBossServerUIPlugin.log("No wf plugin " + plugin, null);
 			}
 	    } catch( Exception e ) {
